backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.16.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-core==1.36.0
azure-eventhub==5.15.1
azure-eventhub-checkpointstoreblob==1.2.0
azure-iot-device==2.14.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
eventlet==0.40.3
faiss-cpu==1.10.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.17.0
Flask==3.1.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
langsmith==0.3.11
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.5.0
networkx==3.4.2
numpy==2.2.3
orjson==3.10.15
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
patsy==1.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==11.0.0
propcache==0.3.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.10.6
pydantic-extra-types==2.11.0
pydantic-settings==2.8.1
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.2
PyMySQL==1.1.1
pyodbc==5.3.0
pyparsing==3.2.0
pypdfium2==5.3.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-engineio==4.12.3
python-multipart==0.0.22
python-socketio==5.14.3
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
requests-unixsocket2==1.0.1
rich==14.3.1
rich-toolkit==0.17.1
rignore==0.7.6
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.15.2
sentence-transformers==3.4.1
sentry-sdk==2.51.0
setuptools==75.8.2
shellingham==1.5.4
shiboken6==6.8.2.1
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.50.0
statsmodels==0.14.5
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
ttkbootstrap==1.12.0
typer==0.21.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.5.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.3
wsproto==1.2.0
xgboost==3.1.2
yarl==1.18.3
zstandard==0.23.0
